requirements.txt
asgiref==3.8.1
Django==5.1.3
django-enum==2.0.2
django-filter==24.3
django-mysql==4.15.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.3
pytest-django==4.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
wheel==0.45.1
